test(NNS): isolate upgrade_everything_test to its own bazel target
